Added full book demo including use of experimental Chrome offscreen tabs extension API.
Initial implementation of this extension API (in the process of being
checked in) still has a couple of bugs preventing this demo from
working completely; in particular, pages using the accelerated
compositor (WebGL and HTML5 video, i.e., YouTube) don't render
correctly.

Once this API has been committed, run chrome with the command line
argument --enable-experimental-extension-apis , and load the extension
as an unpacked extension. For the moment, if you also run with
--disable-accelerated-compositing, the WebGL pages will render
correctly, though the overall demo will run more slowly.

It's also necessary to opt in to YouTube's HTML5 trial
(http://youtube.com/html5) for the video to work.